+++ src/main/java/lotto/Application.java
@@ -1,5 +1,8 @@
 package lotto;
 
+import lotto.view.Input;
+import lotto.view.InputValidator;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
