+++ src/main/java/lotto/view/InputValidator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.view;
 
 import java.util.*;
